add log csv elements tests
